feat(110): Consolidates CharacterSettingsModal into CharacterVitalsEditModal
